inxdex array ex 5

+++ src/main/java/rtp/HeaderFormat.java
@@ -37,9 +37,9 @@ public abstract class HeaderFormat {
 	
 	protected List<Integer> contribSrc;
 	
-	protected short extHeaderID = -1;
+	protected int extHeaderID = -1;
 	
-	protected short extLength = -1;
+	protected int extLength = -1;
 	
 	protected List<Integer> extensions;
 	
@@ -309,9 +309,9 @@ public abstract class HeaderFormat {
 	 */
 	protected void readExtensions() throws RTPHeaderException{
 		if (this.extentionX == 1) {
-			this.extHeaderID = (short) ((this.header[(1+RTPHeader.headerminlength+this.cc)] * 256^2)+ ((this.header[(RTPHeader.headerminlength+this.cc)])));
+			this.extHeaderID = ((this.header[(1+RTPHeader.headerminlength+this.cc)] * 256^2)+ ((this.header[(RTPHeader.headerminlength+this.cc)])));
 			RTPServerLog.log("Extension header ID in idx "+(1+RTPHeader.headerminlength+this.cc)+ ","+(RTPHeader.headerminlength+this.cc)+" is " + this.extHeaderID);
-			this.extLength = (short)((this.header[(3+RTPHeader.headerminlength+this.cc)] * 256^2)+ ((this.header[(2+RTPHeader.headerminlength+this.cc)])));
+			this.extLength = ((this.header[(3+RTPHeader.headerminlength+this.cc)] * 256^2)+ ((this.header[(2+RTPHeader.headerminlength+this.cc)])));
 			RTPServerLog.log("Extension Length in idx "+(3+RTPHeader.headerminlength+this.cc)+ ","+(2+RTPHeader.headerminlength + this.cc)+" is " + this.extLength);
 			
 			this.extensions = new ArrayList<Integer>();
@@ -443,14 +443,14 @@ public abstract class HeaderFormat {
 	/**
 	* Extension ID
 	*/
-	public short getExtHeaderID() {
+	public int getExtHeaderID() {
 		return extHeaderID;
 	}
 
 	/**
 	* Extension length
 	*/
-	public short getExtLength() {
+	public int getExtLength() {
 		return extLength;
 	}
